backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
earthengine-api==0.1.374
requests==2.31.0
matplotlib==3.7.2
pandas==2.0.3
numpy==1.24.3
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.0
aiosqlite==0.19.0
pydantic==2.5.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
PyJWT==2.8.0